+++ lib/src/experimental/auto_translate.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-// import 'dart:io';
 import 'package:autotranslator_widget/src/Translations.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -47,40 +46,7 @@ class LanguageProvider2 with ChangeNotifier {
 
   /// listen to connectivity changes ONCE.
   void _initConnectivityListener() {
-    // _connectivitySubscription =
-    //     this.connectivity.onConnectivityChanged.listen((results) {
-    //   final online = results == ConnectivityResult.mobile ||
-    //       results == ConnectivityResult.wifi ||
-    //       results == ConnectivityResult.ethernet;
-    //     print(' Connectivity changed: ${online ? 'Online' : 'Offline'}');
-    //   if (online != _isOnline) {
-    //     _isOnline = online;
-    //     notifyListeners();
-    //   }
-    // });
-  //   _connectivitySubscription =
-  //     connectivity.onConnectivityChanged.listen((results) async {
-  //   final connected =
-  //       results == ConnectivityResult.mobile ||
-  //       results == ConnectivityResult.wifi ||
-  //       results == ConnectivityResult.ethernet;
-
-  //     print("Device Connectivity changed: ${connected ? 'Online' : 'Offline'}");
-  //   if (!connected) {
-  //     // Interface may be down → report offline
-  //     _isOnline = false;
-  //     notifyListeners();
-  //     return;
-  //   }
-
-  //   // Check actual internet access
-  //   final internet = await _hasInternet();
-
-  //   if (_isOnline != internet) {
-  //     _isOnline = internet;
-  //     notifyListeners();
-  //   }
-  // });
+   
   InternetConnection().onStatusChange.listen((status) {
   _isOnline = status == InternetStatus.connected;
   print(' Internet connection status changed: ${_isOnline ? 'Online' : 'Offline'}');